Add a Numba power-on-self-test script and use in CI.
This replaces the much broader "test of runtests" check performed
in CI with a more simple check aiming to ensure that "runtests"
works and Numba can correctly compile and execute at least one
thing. The intent of this change is to drastically reduce the run
time spent on this check.