CUDA Simplify get_cudalib() and test()
The platform argument is always `None` or the current platform,
resulting in the same (default) behaviour as if it were not supplied, so
it is superfluous.

Testing the libraries should always print the path and not the basename,
because omitting the full path only makes it harder to troubleshoot
issues.