Report the mapped libcuda.so in CUDA libs test
Various driver-related issues have been reported by WSL2 users, and it
is almost always due to a Linux (i.e. not- WSL2) driver being installed
in a WSL2 system.

This commit modifies the test function so that on Linux, it also
provides the absolute location of the driver, which indicates its
version in the soname (e.g. "libcuda.so.530.30.02"). This can then be
used to look up whether the driver was intended for "native" Linux.

The simplest solution for this seems to be to use the maps file for the
current process to look up which libraries are loaded - I could not find
a way through the ctypes API to get the location of the loaded library.

With this change, the output of the libs test function now looks like
(e.g. on my system):

```
Finding driver from candidates: libcuda.so, libcuda.so.1, /usr/lib/libcuda.so, /usr/lib/libcuda.so.1, /usr/lib64/libcuda.so, /usr/lib64/libcuda.so.1...
Using loader <class 'ctypes.CDLL'>
	trying to load driver...	ok, loaded from libcuda.so
	Mapped libcuda.so paths:
		/usr/lib/x86_64-linux-gnu/libcuda.so.530.30.02
Finding nvvm from Conda environment
	located at /home/gmarkall/mambaforge/envs/numbadev/lib/libnvvm.so.4.0.0
	trying to open library...	ok
Finding cudart from Conda environment
	located at /home/gmarkall/mambaforge/envs/numbadev/lib/libcudart.so.11.8.89
	trying to open library...	ok
Finding cudadevrt from Conda environment
	located at /home/gmarkall/mambaforge/envs/numbadev/lib/libcudadevrt.a
Finding libdevice from Conda environment
	trying to open library...	ok
```

Note the new "Mapped libcuda.so paths" section. Since this code is also
used for the Numba sysinfo tool, similar output is also seen from `numba
-s`.