Run tests that require NVRTC even with ctypes binding
This commit is authored in Michael Collison's name to preserve
attribution for his work (though it has been aggregated from changes in
PR #8893).

Tests of float16 division need to be skipped with NVVM 3.4 - this was
never working due to an NVVM 3.4 code generation bug, but was not
noticed before. It became apparent once tests started running in CI on
the old toolkit versions that include NVVM 3.4.