Add an `@overload` API checking function and associated test.
This adds a function that checks Numba's internal use of
`@overload` to make sure that the signature of the overload
implementation and the signature of the API function being
overloaded match in terms of argument names. The function is
used in a unit test that walks the Numba registries to check
all declared `@overloads`. The checking function itself is
imperfect, external APIs change and also have issues like
missing signatures, as a result the unit test is "best effort"
only.

Also note that the checking function can be used to report
where violations occur when run in "debug" mode, this makes it
easier to locate and fix such violations.