Add support for a 'max' level to NUMBA_OPT environment variable.
This adds support for a new 'max' level to the NUMBA_OPT
environment level. Prior to this values 0, 1, 2 and 3 were
supported, these correspond approximately to the -O{level} flag
available in many command line based compilation tools. Issue
opt=3 both before and after the reference count pruning pass
is run (Numba historically had this behaviour as default). This
patch re-enables this behaviour and highlights the caveats in the
docs (performance may not improve, compilation time may increase
and there be no performance gain, but the opposite is also
possible, the user can now choose!).

Closes #8172