CUDA: Vendor half-precision floating point headers
In order to ensure the half-precision floating point headers are
available in all installations, we vendor them from the CUDA toolkit
11.2 (chosen as it is the oldest supported toolkit version, and
therefore expected to be compatible with all supported NVRTC versions).

These headers are redistributable as per the CUDA EULA, and explicitly
mentioned in Attachment A at
https://docs.nvidia.com/cuda/archive/11.2.2/eula/index.html#attachment-a
under the "CUDA Half Precision Headers" component.