Fix uncaught exception in `find_file()`
`find_file()` assumed that all folders it was asked to search actually
exist - this may not be the case, and is exposed now that we search more
locations for CUDA packages. To prevent uncaught exceptions propagating
outward (and thereby crashing `numba -s`, `cuda.cudadrv.libs.test()`,
etc.), we catch `FileNotFoundError` when listing directories, and
continue with the next dir when this occurs.