enable preserving backlinks
The file `CHANGE_LOG` is included as `source/release-notes.rst` instead
of `source/release/CHANGE_LOG.rst. There is a page reachable from the
`index.rst` called `source/release-notes-overview.rst` that pints to
the towncrier notes and the old notes.

This means that, when rendered on Read The Docs the links of the form:

https://numba.readthedocs.io/en/stable/release-notes.html#version-0-57-1-21-june-2023

will continue to work and not be in:

https://numba.readthedocs.io/en/stable/release/CHANGE_LOG.html#version-0-57-1-21-june-2023

And so any existing links to specific release notes will continue to
work.