Support passing arbitrary flags to NVVM
Maintaining a list of flags that we know about for NVVM and validating
them is extra work to keep in sync with the CUDA documentation (our
current docstring for the flags is out-of-date at present).

This PR removes this maintenance burden and allows a bit more
flexibility for future changes by allowing arbitrary flags to be passed
through, transforming them in the same manner as before but without a
restriction on what the flag names are.

Some flags may not take a parameter - `-gen-lto` is an example of this.
To support this case, options can be passed with a value of `None`.
Since this expands slighty on the functionality, a new test cases is
added with this particular flag, that validates it was passed correctly
based on the NVVM output.