CUDA: Use i64 for cude.grid[size] functions
i32 is too small to hold the maximum potential x dimension, so it needs
to use i64 instead.

Technically it would be possible to return a tuple of (i64, i32, i32)
for 2-/3-dimensional grids because the maximum dimensionality of these
is (65535 * <1024 or 64>) but I'm not sure the additional complexity is
worth doing for the potential register /artithmetic intensity saving -
I'd be keener to recommend manual computation of indices using the
thread and block indices / dimensions to anyone aggressively optimizing
for register usage at this point.