CUDA: Add a used list entry for kernels
The compiler or linker are at liberty to drop kernels if they are not
explicitly added to the `@"llvm.used"` global variable. This has not
been an issue so far because NVVM's behaviour has always been to
preserve kernels; however, compiling / linking with nvjitlink causes
anything not referenced in the used list (directly or indirectly) to be
dropped.

When setting a function as a CUDA kernel, a used list is now created and
the kernel is added to it - this assumes one kernel per module, which is
true for Numba. Testing ensures that we correctly construct the used
list, and the existing test IR in `test_nvvm_driver` is updated to
contain a used list.