CUDA compilation: allow specification of C ABI name
Name clashes between different compiled variants of a function will be
frequent and inconvenient if their names all match the Python function
name - this commit allows the name to be specified.

The name is specified under a key in an `abi_info` dict rather than just
a kwarg for the name - this is because future ABIs might require other
options / info, and having a dict that contains all the configuration
for the ABI will allow flexibility in future without revising the API
for compilation again.