CUDA compiler: Add path to compile to LTO-IR
We create a more generic function, `numba.cuda.compile()`, that provides
similar functionality to `compile_ptx()`, but allowing the choice of PTX
or LTO-IR output. This function defaults to the C ABI rather than the
Numba one, as this is expected to be more convenient for most use cases.
We also add a variant to target the current device.

The original `compile_ptx()` and variant for the current device are left
in to support existing use cases that use them and expect generated
code to use the Numba ABI.