[py312] [controlflow] Fix `literal_unroll` label collisions.
This fixes an issue with the `literal_unroll` compiler pass where
an error in the analysis lead to the special case of nested
unrollable loops with the outer loop induction variable
unreferenced, writing to the same key in an internal dictionary
as another legitimate unrollable loop. The result of this collision
was that more unrolling was undertaken than permissable and was done
so incorrectly, which later lead to type inference failures (and
probably invalid IR).

On further consideration, it was likely a mistake to have originally
considered this case as supported (see notes in
`TestMixedTupleUnroll.test_22`) largely because the `literal_unroll`
had no effect and could trivially be replaced with a `range(len())`
based loop. As a result the fix in this patch comprises:

1. Enforcing the condition of "nested `literal_unroll` loops are
   unsupported".
2. Some refactoring to accommodate the above.
3. Making the `test_22` mentioned above into an expected unsupported
   case.