CUDA tests: fix `skip_if_cudadevrt_missing()`
Commit 2bd86f9609 changed `check_static_cudalib()` to expect a path
instead of a library name. However, `cudadevrt_missing()` was not
updated to pass the path, so it always thought `libcudadevrt.a` was
missing, and therefore any tests relying on it (e.g. cooperative groups
tests) have always been skipped since then.

This commit fixes the implementation to pass a path to
`check_static_cudalib()`, and therefore reinstates the CGs tests.