CUDA: Switch cooperative groups to use overloads
This is a non-functional change that only moves the CGs API
implementation up to the high-level extension API. Note that CGs were
never defined with a data model, so it's a little confusing to see
what's going on when the grid group is passed to the `sync()` method,
but it's an `i64` that is passed around opaquely.

Type hints are added on the public API to help IDEs / autocompleting
REPLs.