Add `@box` impl for UndefVar type that raises.
This fixes the new loop-lifting "object mode fall-back"
implementation by making it so that trying to box an UndefVar type
raises a TypingError, i.e. it makes the error compile time opposed
to runtime. This lets the loop-lifting go down the "object mode
fall-back" route and lets compilation succeed in object mode. Prior
to this, a nopython mode loop-lift would succeed but the compiled
implementation would have contain a RuntimeError that will be
unconditionally raised.

The test that failed which prompted this change is:

`numba.tests.test_object_mode.TestObjectMode.test_expr_undef`