Prevent too-new versions of CUDA runtime and NVRTC installing
When CUDA 11.5 is used in gpuCI, cuda-python 11.8 is installed - this is
to ensure the latest version of the cuda-python package is used, and it
is compatible with earlier toolkit versions. It depends on cuda-cudart
and cuda-nvrtc, which are installed alongside it. When cuda-python 11.8
is installed the 11.8 version of these other packages is also picked by
default, so we end up with a mix of versions of toolkit components,
leading to incorrect behaviour - in particular, Numba determines from
the runtime version that CC 8.9 is supported, even though the toolkit
11.5 does not support it.

This commit aims to rectify the issue by choosing to specifically
install the 11.5 versions of toolkit components alongside cuda-python
11.8