Don't attempt to register overloads that aren't for this target
This moves the target check from inside the `_OverloadMethodTemplate` (a
check which was curiously missing from the same place in
`_OverloadAttributeTemplate` - how did that ever work, e.g. for
`cuda.warpsize`, which is an overloaded attribute specific to the CUDA
target?) to the `BaseContext.install_registry()` method.

This could be seen as a more appropriate place for the check - to filter
out methods / attributes that aren't for the current target instead of
attempting to register them then silently doing nothing when a target
mismatch is detected.