`BaseContext`: Only attempt to register overloads that match the current target
This moves the target check from inside the `_OverloadMethodTemplate` to
the `BaseContext.install_registry()` method. The effect of this change
is that instead of attempting to register all overloaded method
templates with all targets and then silently failing inside the
registration, we only attempt to register overloaded methods that are
appropriate for the target.

The `_OverloadAttributeTemplate` never contained this check, which
appears curious at first sight - however, there is a related bug in
overloading of attributes that causes their metadata to be lost, and
therefore all attributes get registered for all targets regardless of
whether they are for the current target or not. This error will be
addressed in the following commit.