Stub support for LTO for `@cuda.jit` functions
Previous commits added support for compiling Python functions to CUDA
LTO-IR via the compilation interfaces. This commit adds stub code for
supporting compilation of `@cuda.jit`-decorated functions to LTO-IR. The
only functional change, unused in Numba at present, is that if the
linker has LTO enabled, the CUDA codegen uses NVVM to generate LTO-IR
instead of PTX, and passes that to the linker.

The `lto` attribute is added linker classes in
`numba.cuda.cudadrv.driver` - this is always `False` for the built-in
linkers, but a linker from pynvjitlink (or any other external linker, in
theory) could set it to `True` to signal that LTO is enabled.

Some tests must be skipped if LTO is enabled, because it becomes
difficult to use the functionality they test when LTO is enabled:

- Some inspect the PTX, which is difficult to do when LTO-IR is
  generated instead.
- Others check for exceptions, but the exception flags get optimized
  away by LTO because Numba fails to add them to the used list (See
  #9526).