Ignore registrations with invalid target during typing context refresh
When typing contexts are refreshed, there may be registrations for
invalid targets pending (e.g. from other tests that exercise error
handling in the presence of invalid user code). These registrations need
to be ignored - see the comments in the code for further detail.

Attempting to compile a call to an overload with an invalid target now
causes a `NonexistentTargetError` to be raised, so the target extension
tests are updated accordingly.