Fix cProfile for Python < 3.11 when numba raises.
If a Numba function is being run under cProfile and the
function raises an exception, the dispatcher must handle
this correctly. It is not valid to call
`PyFrame_FastToLocals` without saving and restoring the
exception state across the call as it can clear the
exception state itself. This patch fixes this problem and
also stops calling `PyFrame_LocalsToFast` in Python 3.11
where the "whats-new" docs claim that the frames are now
looked after by the virutal machine.