Workaround potential CPython bug.
This works around a potential bug in CPython where the state
associated with a monitoring tool can "leak" in an unexpected
manner.

This patch also removes the use of the functools cache on the
generator for test functions, this to eliminate any potential
state being stored on the code objects associated with the
test functions.