Add CUDA target implementation to sysinfo and module
In preparation for a time when the CUDA target is maintained out of the
core Numba tree, this adds the name of the target to the Numba sysinfo.
The name for the existing in-tree `numba.cuda` module is "Built-in",
whereas an out-of-tree implementation might be called "NVIDIA", for
example.

If there is an out-of-tree target in use that does not provide its name,
we display `'?'` as the implementation name.

The name is also available programmatically under
`numba.cuda.implementation`.

This should aid in determining the source of problems / where to direct
users for further support when an external target is in use.