activate compilers for linux-aarch64
Historically the Numba conda recipe has not requested to install
compilers on linux-aarch64 and relied on the system compilers instead.
The toolchain has now matured.

With the change to using the docker container `continuumio/miniconda3`
for Anaconda internal testing of Numba, we will need to install the
correct compilers and change the recipe such that they are used.

This is a followup from a similar change needed for llvmlite:

https://github.com/numba/llvmlite/pull/1073