Fix issue with IR mutation in label renaming.
This fixes an issue in the "objectmode fallback" emulation pipeline.
Prior to this fix, the inline closure call would run as part of a
"nopython" pipeline and the label renamer would mutate the IR
terminators when renaming in topological order. This mutation would
carry over into a copy of the IR that was used in a subsequent
"objectmode" compilation attempt (assuming the nopython mode attempt
failed). As a result, this invalid IR would not compile as the CFG
cannot be constructed from incorrect labels. The fix is to ensure that
renaming of the terminators is done via constructing a replacement node
opposed to direct mutation of an existing node.

Fixes #9725