Fix some memory leaks/poor memory behaviour
Issue #9103 showed what appears to be a memory leak
associated with the use of `np.random.Generator` objects.

It turned out that the `np.random.Generator` unboxing code leaked
numerous references, however, fixing these didn't completely stop
the continuous consumption of memory.

Following on from the above, what appeared to be a less drastic
leak was observable. On investigation, it turns out that this is
not a "leak" in that the memory is still addressable and would
eventually be garbage collected. The issue is that in the
dispatcher, the use of the `typeof_pyval` method, as occurs in the
case of a type which cannot be fingerprinted, appends the result
of computing the type to a list so as to keep a reference alive.
This append is unconditional, therefore every dispatch adds to the
list. The fix is to make this a `set` such that references are held,
but only once, and memory use does not grow with each invocation of
the dispatcher with a non-fingerprintable type.

Fixes #9733