Fix PyRef leaks in unboxing of np.random.Generator
Issue #9103 shows what appears to be a memory leak
associated with the use of `np.random.Generator` objects. The unboxed
`Generator` object has a `bit_generator` field which refers to a struct
formed from an unboxed `np.random.BitGenerator`. It is the unboxing of the
`BitGenerator` instances that are leaking numerous Python reference
(the leaking code is called 3x per unbox). This patch fixes those leaks.

Fixes #9103