Fix miscompile in branch pruning of SSA form IR.
Issue #9742 demonstrates a problem in SSA form IR where a branch
predicate that is a binop expression with constant arguments
was erroneously being rewritten as a constant 1 or 0 depending
on whether the branch would be taken or not. This is incorrect as
the predicate could be referred to by later code and as a result
needs to have the value that is the result of evaluating the binop
expression. This patch fixes the problem and adds test cases to
demonstrate.

Fixes #9742