Prevent bools from if ops being interpreted as user variables
The convention for user variables is that they don't begin with "$".
Bools from if ops violate this convention by using the name "bool<X>".

The result of them being seen as user variables is that they appear in
the debug info, which we don't want (e.g. `info locals` will show a lot
of `bool<X>` names that will be unfamiliar to the user).

This commit prepends a $ to their name in order to keep their names
hidden from debug info.