Document and test `as_numba_type`'s role in instance checks
Originally `as_numba_type` was added to facilitate the inference of
Numba types from Python type annotations on jitclass fields.

However, since #7244 it also plays a role in `isinstance` checks, as it
is used to determine the type of the `classinfo` argument to the
`isinstance()` function.

This PR adds a test demonstrating its use (and elucidating some of the
pre-requisites for the instance check to work) and updates the
documentation to explain its use in instance checks.

There are mentions of `as_numba_type()` in a couple of other files that
are unmodified:

- `source/extending/low-level.rst`
- `source/user/jitclass.rst`

The wording in these sections already appeared to be appropriate so they
are left unchanged.