Track when all the types are unsigned
Was running into an issue that if there was a single unsigned type and no floats, calling `_arange_dtype` threw an error since max was trying to take the max of a signed and unsigned type which isn't allowed. So track whether all the types are unsigned, and set the types to `np.uint64`.

Also, instead of casting `step` to `dtype` which might be None, cast to `true_dtype`, which is guaranteed to not be None.