Fix #10281: generator expression used as args
`_call_function_ex_replace_args_large()` expects that the `vararg_stmt`
is an `ir.Assign` whose value is an `ir.Var`. This is not the case when
the arguments come from a generator expression, and it is an `ir.Expr`
instead. Since this is unsupported anyway (and will report an error
later in the pipeline) it seems that a pragmatic fix is to inform the
user at this point that using a generator expression as arguments is
unsupported; fixing up the function to correctly handle assigning an
`ir.Expr` seems like more work that will not add value to users, as it
will only result in a less-clear error message ("The use of yield in a
closure is unsupported") being given to the user anyway.